+++ core/src/main/java/org/apache/struts/action/ActionFormBean.java
@@ -43,7 +43,7 @@ public class ActionFormBean extends FormBeanConfig {
     private static final long serialVersionUID = -6870747200275478095L;
 
     /**
-     * <p>Construct an instance with default vaslues.</p>
+     * <p>Construct an instance with default values.</p>
      */
     public ActionFormBean() {
         super();